logexec: Perl-CS done

+++ services/logexec/AdvancedFiltering/DB/submissiondb.pm
@@ -14,6 +14,7 @@ sub add_message
 	my $realm = shift;
 	my $mid = shift;
 	my $sender = shift;
+	die "USAGE: AdvancedFiltering::DB::submissiondb::add_message<realm><mid><sender>" unless defined($realm) && defined($mid) && defined($sender);
 	return add_ldap_object("afUSMTPMessageId=$mid,afUServiceName=submissiondb+afUServiceRealm=$realm,ou=user,o=advancedfiltering",
 		{objectClass => 'afUSMTPMessageOutgoing', afUSMTPMessageSenderMailAddress => $sender, afUSMTPMessageTimeCreated => time});
 }
@@ -22,6 +23,7 @@ sub delete_message
 {
 	my $realm = shift;
 	my $mid = shift;
+	die "USAGE: AdvancedFiltering::DB::submissiondb::delete_message<realm><mid>" unless defined($realm) && defined($mid);
 	return delete_ldap_object("afUSMTPMessageId=$mid,afUServiceName=submissiondb+afUServiceRealm=$realm,ou=user,o=advancedfiltering");
 }
